backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.1.0
cffi==1.15.1
Django==4.2.3
django-ckeditor==6.6.1
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
mysqlclient==2.2.0
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
